Split up commands.py
